Table created

+++ Worldle1/src/project/Country.java
@@ -29,6 +29,7 @@ public class Country
     }
     
     /**
+     * 
      * @return the name of the country
      */
     public String getName()
@@ -77,7 +78,7 @@ public class Country
         return "Country: " + name + ", Latitude: " + latitude + ", Longitude: " + longitude + ", Code: " + code;
     }
 
-    public double getDistance(Country other)
+    public int getDistance(Country other)
     {
         double lat1 = this.getLatitude();
         double lon1 = this.getLongitude();
@@ -88,11 +89,12 @@ public class Country
         double dLon = Math.toRadians(lon2 - lon1);
         double a = Math.sin(dLat/2) * Math.sin(dLat/2) + Math.cos(Math.toRadians(lat1)) * Math.cos(Math.toRadians(lat2)) * Math.sin(dLon/2) * Math.sin(dLon/2);
         double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a));
-        return R * c;
+        return (int) (R * c);
     }
 
     /**
      * @return the direction from this country to another country
+     * from this country to another country
      */
     public String getDirection(Country other)
     {
